requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
django-mysql==4.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mysqlclient==2.2.0
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.4
